Shaping up the codebase
I started creating a new class inside retriever.py,
SinglePhaseRetriever, which will center all the action of the phase
retrieval process. This will serve as the base of the GUI also, in which
I will finally detach the functionality from the interface. To do so,
great changes are needed inside interface.py, but they'll eventually
come.