Major refactoring - simple tests complete
The first step towards a major refactoring has been completed.
The class SimplePhaseRetriever contains now functions to automatically
determine the location of maximum intensity, the centering of the window
and the determination of the function bandwidth. Basic functionality
tests are successfully completed. Now we just need to test whether the
phase retrieval results are consistent or not with the previous
implementation.